+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project.
 ## [Unreleased]
 
 ### Fixed
-- `AnswerPreCheckoutQuery` default parameter.
+- default parameters in `AnswerPreCheckoutQuery`.
 
 
 ## [5.0.0] - 2023-03-11